Add ranks, Convert social media to class and fix issues in Player. Move utils module to a single file.
