+++ hypixelio/_async/client.py
@@ -7,6 +7,7 @@ import typing as t
 import aiohttp
 
 from ..base import BaseClient
+from ..constants import HYPIXEL_API, TIMEOUT
 from ..exceptions import (
     GuildNotFoundError,
     HypixelAPIError,
@@ -28,8 +29,7 @@ from ..models.skyblock import (
     SkyblockActiveAuction, SkyblockBazaar, SkyblockNews, SkyblockProfile, SkyblockUserAuction
 )
 from ..models.watchdog import Watchdog
-from ..utils.constants import HYPIXEL_API, TIMEOUT
-from ..utils.url import form_url
+from ..utils import form_url
 
 
 class AsyncClient(BaseClient):
