+++ hypixelio/constants.py
