+++ hypixelio/models/guild.py
@@ -1,4 +1,4 @@
-from hypixelio.utils.time import unix_time_to_datetime
+from hypixelio.utils import unix_time_to_datetime
 
 
 class Guild:
